+++ src/Header/CacheControl.php
@@ -3,7 +3,16 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
-use Innmind\Http\Header;
+use Innmind\Http\{
+    Header,
+    Header\CacheControl\Directive,
+    Header\CacheControl\MaxAge,
+    Header\CacheControl\MaxStale,
+    Header\CacheControl\MinimumFresh,
+    Header\CacheControl\NoCache,
+    Header\CacheControl\PrivateCache,
+    Header\CacheControl\SharedMaxAge,
+};
 use Innmind\Immutable\Sequence;
 
 /**
@@ -12,7 +21,7 @@ use Innmind\Immutable\Sequence;
 final class CacheControl implements Custom
 {
     /**
-     * @param Sequence<CacheControlValue> $directives
+     * @param Sequence<Directive|MaxAge|MaxStale|MinimumFresh|NoCache|PrivateCache|SharedMaxAge> $directives
      */
     private function __construct(
         private Sequence $directives,
@@ -24,8 +33,8 @@ final class CacheControl implements Custom
      * @no-named-arguments
      */
     public static function of(
-        CacheControlValue $first,
-        CacheControlValue ...$values,
+        Directive|MaxAge|MaxStale|MinimumFresh|NoCache|PrivateCache|SharedMaxAge $first,
+        Directive|MaxAge|MaxStale|MinimumFresh|NoCache|PrivateCache|SharedMaxAge ...$values,
     ): self {
         return new self(Sequence::of($first, ...$values));
     }