+++ tests/Header/CacheControlTest.php
@@ -5,8 +5,8 @@ namespace Tests\Innmind\Http\Header;
 
 use Innmind\Http\{
     Header\CacheControl,
+    Header\CacheControl\Directive,
     Header,
-    Header\CacheControlValue\PublicCache
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -15,7 +15,7 @@ class CacheControlTest extends TestCase
     public function testInterface()
     {
         $h = CacheControl::of(
-            $v = new PublicCache,
+            Directive::public,
         );
 
         $this->assertInstanceOf(Header\Custom::class, $h);